kumo_api_types::egress_path

Struct EgressPathConfig

pub struct EgressPathConfig {

    pub connection_limit: LimitSpec,
    pub additional_connection_limits: OrderMap<String, LimitSpec>,
    pub enable_tls: Tls,
    pub enable_mta_sts: bool,
    pub enable_dane: bool,
    pub enable_pipelining: bool,
    pub enable_rset: bool,
    pub tls_prefer_openssl: bool,
    pub tls_certificate: Option<KeySource>,
    pub tls_private_key: Option<KeySource>,
    pub openssl_cipher_list: Option<String>,
    pub openssl_cipher_suites: Option<String>,
    pub openssl_options: Option<SslOptions>,
    pub rustls_cipher_suites: Vec<SupportedCipherSuite>,
    pub client_timeouts: SmtpClientTimeouts,
    pub system_shutdown_timeout: Option<Duration>,
    pub max_ready: usize,
    pub consecutive_connection_failures_before_delay: usize,
    pub smtp_port: u16,
    pub smtp_auth_plain_username: Option<String>,
    pub smtp_auth_plain_password: Option<KeySource>,
    pub allow_smtp_auth_plain_without_tls: bool,
    pub max_message_rate: Option<ThrottleSpec>,
    pub additional_message_rate_throttles: OrderMap<String, ThrottleSpec>,
    pub source_selection_rate: Option<ThrottleSpec>,
    pub additional_source_selection_rates: OrderMap<String, ThrottleSpec>,
    pub max_connection_rate: Option<ThrottleSpec>,
    pub max_deliveries_per_connection: usize,
    pub max_recipients_per_batch: usize,
    pub prohibited_hosts: CidrSet,
    pub skip_hosts: CidrSet,
    pub ehlo_domain: Option<String>,
    pub aggressive_connection_opening: bool,
    pub refresh_interval: Duration,
    pub refresh_strategy: ConfigRefreshStrategy,
    pub dispatcher_wakeup_strategy: WakeupStrategy,
    pub maintainer_wakeup_strategy: WakeupStrategy,
    pub provider_name: Option<String>,
    pub remember_broken_tls: Option<Duration>,
    pub opportunistic_tls_reconnect_on_failed_handshake: bool,
    pub use_lmtp: bool,
    pub reconnect_strategy: ReconnectStrategy,
    pub readyq_pool_name: Option<String>,
    pub low_memory_reduction_policy: MemoryReductionPolicy,
    pub no_memory_reduction_policy: MemoryReductionPolicy,
    pub try_next_host_on_transport_error: bool,
    pub ignore_8bit_checks: bool,

}

Fields
connection_limit: LimitSpecadditional_connection_limits: OrderMap<String, LimitSpec>enable_tls: Tlsenable_mta_sts: boolenable_dane: boolenable_pipelining: boolenable_rset: booltls_prefer_openssl: booltls_certificate: Option<KeySource>tls_private_key: Option<KeySource>openssl_cipher_list: Option<String>openssl_cipher_suites: Option<String>openssl_options: Option<SslOptions>rustls_cipher_suites: Vec<SupportedCipherSuite>client_timeouts: SmtpClientTimeoutssystem_shutdown_timeout: Option<Duration>

How long to wait for an established session to gracefully close when the system is shutting down. After this period has elapsed, sessions will be aborted.

max_ready: usizeconsecutive_connection_failures_before_delay: usizesmtp_port: u16smtp_auth_plain_username: Option<String>smtp_auth_plain_password: Option<KeySource>allow_smtp_auth_plain_without_tls: boolmax_message_rate: Option<ThrottleSpec>additional_message_rate_throttles: OrderMap<String, ThrottleSpec>source_selection_rate: Option<ThrottleSpec>additional_source_selection_rates: OrderMap<String, ThrottleSpec>max_connection_rate: Option<ThrottleSpec>max_deliveries_per_connection: usizemax_recipients_per_batch: usizeprohibited_hosts: CidrSetskip_hosts: CidrSetehlo_domain: Option<String>aggressive_connection_opening: boolrefresh_interval: Duration

How long to wait between calls to get_egress_path_config for any given ready queue. Making this longer uses fewer resources (in aggregate) but means that it will take longer to detect and adjust to changes in the queue configuration.

refresh_strategy: ConfigRefreshStrategydispatcher_wakeup_strategy: WakeupStrategymaintainer_wakeup_strategy: WakeupStrategyprovider_name: Option<String>

Specify an explicit provider name that should apply to this path. The provider name will be used when computing metrics rollups by provider. If omitted, then

remember_broken_tls: Option<Duration>

If set, a process-local cache will be used to remember if a site has broken TLS for the duration specified. Once encountered, we will pretend that EHLO didn’t advertise STARTTLS on subsequent connection attempts.

opportunistic_tls_reconnect_on_failed_handshake: bool

If true, when a TLS handshake fails and TLS is set to opportunistic, we will re-connect to that host with TLS disabled.

use_lmtp: bool

If true, rather than ESMTP, use the LMTP protocol

reconnect_strategy: ReconnectStrategy

How to behave if we experience either a 421 response, an IO Error, or a timeout while talking to the peer.

readyq_pool_name: Option<String>

Which thread pool to use for processing the ready queue

low_memory_reduction_policy: MemoryReductionPolicy

What to do to newly inserted messages when memory is low

no_memory_reduction_policy: MemoryReductionPolicy

What to do to newly inserted messages when memory is over the soft limit

try_next_host_on_transport_error: bool

If we experience a transport error during SMTP, should we retry the current message on the next host in the connection plan, or immediately consider it a transient failure for that message?

ignore_8bit_checks: bool

If true, don’t check for 8bit compatibility issues during sending, instead, leave it to the remote host to raise an error.