Module kumo_api_types::shaping

Structs

EgressPathConfigValue

Represents an individual EgressPathConfig field name and value. It only allows deserializing from valid EgressPathConfig field + values.

HashableTomlValue

toml::Value is not Hash because it may contain floating point numbers, which are problematic from a Ord and Eq perspective. We’re okay with skirting around that for our purposes here, so we implement our own hashable wrapper around the toml value.

MergedEntry

Regex

Rule

Shaping

Enums

Action

Trigger

Functions

register