kumo_api_types::shaping

Struct HashableTomlValue

pub struct HashableTomlValue { /* private fields */ }

toml::Value is not Hash because it may contain floating point numbers, which are problematic from a Ord and Eq perspective. Weโre okay with skirting around that for our purposes here, so we implement our own hashable wrapper around the toml value.