kumo_api_types

Module shaping

Structs

EgressPathConfigValue

Represents an individual EgressPathConfig field name and value. It only allows deserializing from valid EgressPathConfig field + values.

EgressPathConfigValueUnchecked

This is the type that we actually use to deserialize EgressPathConfigValue items. It doesn’t care about validity; it is used solely to tell serde what shape of data to expect. The validation is performed by the TryFrom impl that is used to convert to the checked form below.

HashableTomlValue

toml::Value is not Hash because it may contain floating point numbers, which are problematic from a Ord and Eq perspective. We’re okay with skirting around that for our purposes here, so we implement our own hashable wrapper around the toml value.

MergedEntry

ProviderEntry

Regex

Rule

Shaping

ShapingMergeOptions

Enums

Action

CheckLevel

ProviderMatch

Trigger

Functions

register